Move string collection tests to their corresponding test classes
