+++ Tests/aweXpect.Tests/ThatTests/Collections/EnumerableShould.ContainTests.cs
@@ -11,6 +11,112 @@ public sealed partial class EnumerableShould
 	public sealed class ContainTests
 	{
 		[Fact]
+		public async Task ForStrings_ShouldCompareCaseSensitive()
+		{
+			string[] sut = ["green", "blue", "yellow"];
+
+			async Task Act()
+				=> await That(sut).Should().Contain("GREEN");
+
+			await That(Act).Should().Throw<XunitException>()
+				.WithMessage("""
+				             Expected sut to
+				             contain "GREEN" at least once,
+				             but it contained it 0 times in [
+				               "green",
+				               "blue",
+				               "yellow"
+				             ]
+				             """);
+		}
+
+		[Fact]
+		public async Task ForStrings_WhenExpectedIsNotPartOfStringEnumerable_ShouldFail()
+		{
+			string[] sut = ["green", "blue", "yellow"];
+
+			async Task Act()
+				=> await That(sut).Should().Contain("red");
+
+			await That(Act).Should().Throw<XunitException>()
+				.WithMessage("""
+				             Expected sut to
+				             contain "red" at least once,
+				             but it contained it 0 times in [
+				               "green",
+				               "blue",
+				               "yellow"
+				             ]
+				             """);
+		}
+
+		[Fact]
+		public async Task ForStrings_WhenExpectedIsPartOfStringEnumerable_ShouldSucceed()
+		{
+			string[] sut = ["green", "blue", "yellow"];
+
+			async Task Act()
+				=> await That(sut).Should().Contain("blue");
+
+			await That(Act).Should().NotThrow();
+		}
+
+		[Fact]
+		public async Task ForStrings_WhenIgnoringCase_ShouldSucceedForCaseSensitiveDifference()
+		{
+			string[] sut = ["green", "blue", "yellow"];
+
+			async Task Act()
+				=> await That(sut).Should().Contain("GREEN").IgnoringCase();
+
+			await That(Act).Should().NotThrow();
+		}
+
+		[Fact]
+		public async Task ForStrings_WithAtLeast_ShouldVerifyCorrectNumberOfTimes()
+		{
+			string[] sut = ["green", "green", "blue", "yellow"];
+
+			async Task Act()
+				=> await That(sut).Should().Contain("green").AtLeast(3.Times());
+
+			await That(Act).Should().Throw<XunitException>()
+				.WithMessage("""
+				             Expected sut to
+				             contain "green" at least 3 times,
+				             but it contained it 2 times in [
+				               "green",
+				               "green",
+				               "blue",
+				               "yellow"
+				             ]
+				             """);
+		}
+
+		[Fact]
+		public async Task ForStrings_WithAtMost_ShouldVerifyCorrectNumberOfTimes()
+		{
+			string[] sut = ["green", "green", "green", "green", "blue", "yellow"];
+
+			async Task Act()
+				=> await That(sut).Should().Contain("green").AtMost(2.Times());
+
+			await That(Act).Should().Throw<XunitException>()
+				.WithMessage("""
+				             Expected sut to
+				             contain "green" at most 2 times,
+				             but it contained it at least 3 times in [
+				               "green",
+				               "green",
+				               "green",
+				               "green",
+				               "blue",
+				               "yellow"
+				             ]
+				             """);
+		}
+		
+		[Fact]
 		public async Task Item_DoesNotEnumerateTwice()
 		{
 			ThrowWhenIteratingTwiceEnumerable subject = new();
