--- Tests/aweXpect.Tests/ThatTests/Collections/StringEnumerableShould.cs
@@ -1,5 +1,0 @@
-﻿namespace aweXpect.Tests.ThatTests.Collections;
-
-public partial class StringEnumerableShould
-{
-}
